Re-use CollectingBatchIterator for NodeStats retrieval
